requirements.txt
streamlit==1.27.0
streamlit-autorefresh==0.1.3
pandas==2.1.0
requests==2.31.0
websocket-client==1.6.0
plotly==5.22.0
numpy==1.26.0